tube_classification/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
ImageHash==4.3.2
ImageIO==2.37.3
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-loader==0.5
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mobile_sam @ git+https://github.com/ChaoningZhang/MobileSAM.git@b01a9ccef3b9e10b099b544efe004d0871802c3b
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
pooch==1.9.0
pydantic==2.12.5
pydantic_core==2.41.5
pyDeprecate==0.5.0
Pygments==2.19.2
PyMatting==1.1.15
pyparsing==3.3.2
pyrealsense2==2.57.7.10387
pytest==9.0.2
python-dateutil==2.9.0.post0
PyWavelets==1.9.0
PyYAML==6.0.3
referencing==0.37.0
rembg==2.0.74
requests==2.33.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
supervision==0.27.0.post2
sympy==1.13.1
tifffile==2026.3.3
timm==1.0.26
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
win32_setctime==1.2.0
